+++ src/openboardview/main_dx9.cpp
@@ -9,7 +9,7 @@
  */
 #include "BoardView.h"
 
-#include "imgui_impl_dx9.h"
+#include "Renderers/imgui_impl_dx9.h"
 #include <d3d9.h>
 #define DIRECTINPUT_VERSION 0x0800
 #include "confparse.h"
